united all integration tests
