--- nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/java/org/apache/nifi/processors/elasticsearch/docker/DockerComposeContainerType.java
@@ -1,6 +1,0 @@
-package org.apache.nifi.processors.elasticsearch.docker;
-
-public enum DockerComposeContainerType {
-    BASIC,
-    PROXY
-}
