+++ nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/resources/docker-compose.yml
@@ -1,12 +1,29 @@
 version: '2.2'
 services:
-  es01:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.0
+  squid:
+    image: sameersbn/squid:3.5.27-2
+    networks:
+      - mynetwork
+    ports:
+      - 3228:3228
+    volumes:
+      - ./squid.conf:/etc/squid/squid.conf
+    restart: always
+  squid_auth:
+    image: sameersbn/squid:3.5.27-2
+    networks:
+      - mynetwork
+    ports:
+      - 3328:3328
+    volumes:
+      - ./squid-auth.conf:/etc/squid/squid.conf
+  es01_proxy:
+    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.1
     environment:
-      - node.name=es01
-      - cluster.name=es-docker-cluster
-      - discovery.seed_hosts=es02,es03
-      - cluster.initial_master_nodes=es01,es02,es03
+      - node.name=es01_proxy
+      - cluster.name=es-docker-cluster-proxy
+      - discovery.seed_hosts=es02_proxy,es03_proxy
+      - cluster.initial_master_nodes=es01_proxy,es02_proxy,es03_proxy
       - bootstrap.memory_lock=true
       - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
     ulimits:
@@ -15,44 +32,52 @@ services:
         hard: -1
     volumes:
       - data01:/usr/share/elasticsearch/data
-
-  es02:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.0
+    ports:
+      - 9200:9200
+    networks:
+      mynetwork:
+        ipv4_address: 172.18.0.2
+  es02_proxy:
+    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.1
+    depends_on:
+      es01_proxy:
+        condition: service_started
     environment:
-      - node.name=es02
-      - cluster.name=es-docker-cluster
-      - discovery.seed_hosts=es01,es03
-      - cluster.initial_master_nodes=es01,es02,es03
+      - node.name=es02_proxy
+      - cluster.name=es-docker-cluster-proxy
+      - discovery.seed_hosts=es01_proxy,es03_proxy
+      - cluster.initial_master_nodes=es01_proxy,es02_proxy,es03_proxy
       - bootstrap.memory_lock=true
       - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
-    depends_on:
-      es01:
-        condition: service_started
     ulimits:
       memlock:
         soft: -1
         hard: -1
     volumes:
       - data02:/usr/share/elasticsearch/data
-
-  es03:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.0
+    networks:
+      - mynetwork
+  es03_proxy:
+    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.1
+    depends_on:
+      es01_proxy:
+        condition: service_started
     environment:
-      - node.name=es03
-      - cluster.name=es-docker-cluster
-      - discovery.seed_hosts=es01,es02
-      - cluster.initial_master_nodes=es01,es02,es03
+      - node.name=es03_proxy
+      - cluster.name=es-docker-cluster-proxy
+      - discovery.seed_hosts=es01_proxy,es02_proxy
+      - cluster.initial_master_nodes=es01_proxy,es02_proxy,es03_proxy
       - bootstrap.memory_lock=true
       - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
-    depends_on:
-      es01:
-        condition: service_started
     ulimits:
       memlock:
         soft: -1
         hard: -1
     volumes:
       - data03:/usr/share/elasticsearch/data
+    networks:
+      - mynetwork
+
 
 volumes:
   data01:
@@ -61,4 +86,11 @@ volumes:
     driver: local
   data03:
     driver: local
-    
+
+networks:
+  mynetwork:
+    driver: bridge
+    ipam:
+      driver: default
+      config:
+        - subnet: 172.18.0.0/16
