--- nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/resources/proxy-squid/docker-compose-proxy.yml
@@ -1,96 +1,0 @@
-version: '2.2'
-services:
-  squid:
-    image: sameersbn/squid:3.5.27-2
-    networks:
-      - mynetwork
-    ports:
-      - 3228:3228
-    volumes:
-      - ./squid.conf:/etc/squid/squid.conf
-    restart: always
-  squid_auth:
-    image: sameersbn/squid:3.5.27-2
-    networks:
-      - mynetwork
-    ports:
-      - 3328:3328
-    volumes:
-      - ./squid-auth.conf:/etc/squid/squid.conf
-  es01:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.1
-    environment:
-      - node.name=es01
-      - cluster.name=es-docker-cluster-proxy
-      - discovery.seed_hosts=es02,es03
-      - cluster.initial_master_nodes=es01,es02,es03
-      - bootstrap.memory_lock=true
-      - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-    volumes:
-      - data01:/usr/share/elasticsearch/data
-    networks:
-      mynetwork:
-        ipv4_address: 172.18.0.2
-
-  es02:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.1
-    depends_on:
-      es01:
-        condition: service_started
-    environment:
-      - node.name=es02
-      - cluster.name=es-docker-cluster-proxy
-      - discovery.seed_hosts=es01,es03
-      - cluster.initial_master_nodes=es01,es02,es03
-      - bootstrap.memory_lock=true
-      - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-    volumes:
-      - data02:/usr/share/elasticsearch/data
-    networks:
-      - mynetwork
-
-  es03:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.13.1
-    depends_on:
-      es01:
-        condition: service_started
-    environment:
-      - node.name=es03
-      - cluster.name=es-docker-cluster-proxy
-      - discovery.seed_hosts=es01,es02
-      - cluster.initial_master_nodes=es01,es02,es03
-      - bootstrap.memory_lock=true
-      - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-    volumes:
-      - data03:/usr/share/elasticsearch/data
-    networks:
-      - mynetwork
-
-
-volumes:
-  data01:
-    driver: local
-  data02:
-    driver: local
-  data03:
-    driver: local
-
-networks:
-  mynetwork:
-    driver: bridge
-    ipam:
-      driver: default
-      config:
-        - subnet: 172.18.0.0/16
